Swift 5.5: implicitly converts between CGFloat and Double in most places

+++ DiaBLE/Views/Graph.swift
@@ -24,10 +24,10 @@ struct Graph: View {
             // Glucose range rect in the background
             GeometryReader { geometry in
                 Path() { path in
-                    let width  = Double(geometry.size.width) - 60.0
-                    let height = Double(geometry.size.height)
-                    let yScale = (height - 20.0) / yMax()
-                    path.addRect(CGRect(x: 1.0 + 30.0, y: height - settings.targetHigh * yScale + 1.0, width: width - 2.0, height: (settings.targetHigh - settings.targetLow) * yScale - 1.0))
+                    let width  = geometry.size.width - 60
+                    let height = geometry.size.height
+                    let yScale = (height - 20) / yMax()
+                    path.addRect(CGRect(x: 1 + 30, y: height - settings.targetHigh * yScale + 1.0, width: width - 2, height: (settings.targetHigh - settings.targetLow) * yScale - 1))
                 }.fill(Color.green).opacity(0.15)
             }
 
@@ -35,15 +35,15 @@ struct Graph: View {
             GeometryReader { geometry in
                 ZStack {
                     Text("\(settings.targetHigh.units)")
-                        .position(x: CGFloat(Double(geometry.size.width) - 15.0), y: CGFloat(Double(geometry.size.height) - (Double(geometry.size.height) - 20.0) / yMax() * settings.targetHigh))
+                        .position(x: geometry.size.width - 15, y: geometry.size.height - (geometry.size.height - 20) / yMax() * settings.targetHigh)
                     Text("\(settings.targetLow.units)")
-                        .position(x: CGFloat(Double(geometry.size.width) - 15.0), y: CGFloat(Double(geometry.size.height) - (Double(geometry.size.height) - 20.0) / yMax() * settings.targetLow))
+                        .position(x: geometry.size.width - 15, y: geometry.size.height - (geometry.size.height - 20) / yMax() * settings.targetLow)
                     let count = history.rawValues.count
                     if count > 0 {
                         let hours = count / 4
                         let minutes = count % 4 * 15
                         Text((hours > 0 ? "\(hours) h\n" : "") + (minutes != 0 ? "\(minutes) min" : ""))
-                            .position(x: 5.0, y: CGFloat(Double(geometry.size.height) - (Double(geometry.size.height)) / 2.0))
+                            .position(x: 5, y: geometry.size.height - geometry.size.height / 2)
                     }
                 }.font(.footnote).foregroundColor(.gray)
             }
@@ -53,13 +53,13 @@ struct Graph: View {
                 let count = history.rawValues.count
                 if count > 0 {
                     Path() { path in
-                        let width  = Double(geometry.size.width) - 60.0
-                        let height = Double(geometry.size.height)
+                        let width  = geometry.size.width - 60
+                        let height = geometry.size.height
                         let v = history.rawValues.map{$0.value}
-                        let yScale = (height - 20.0) / yMax()
+                        let yScale = (height - 20) / yMax()
                         let xScale = width / Double(count - 1)
                         var startingVoid = v[count - 1] < 1 ? true : false
-                        if startingVoid == false { path.move(to: .init(x: 0.0 + 30.0, y: height - Double(v[count - 1]) * yScale)) }
+                        if startingVoid == false { path.move(to: .init(x: 0 + 30, y: height - Double(v[count - 1]) * yScale)) }
                         for i in 1 ..< count {
                             if v[count - i - 1] > 0 {
                                 let point = CGPoint(x: Double(i) * xScale + 30.0, y: height - Double(v[count - i - 1]) * yScale)
@@ -80,13 +80,13 @@ struct Graph: View {
                 let count = history.factoryValues.count
                 if count > 0 {
                     Path() { path in
-                        let width  = Double(geometry.size.width) - 60.0
-                        let height = Double(geometry.size.height)
+                        let width  = geometry.size.width - 60
+                        let height = geometry.size.height
                         let v = history.factoryValues.map{$0.value}
-                        let yScale = (height - 20.0) / yMax()
+                        let yScale = (height - 20) / yMax()
                         let xScale = width / Double(count - 1)
                         var startingVoid = v[count - 1] < 1 ? true : false
-                        if startingVoid == false { path.move(to: .init(x: 0.0 + 30.0, y: height - Double(v[count - 1]) * yScale)) }
+                        if startingVoid == false { path.move(to: .init(x: 0 + 30, y: height - Double(v[count - 1]) * yScale)) }
                         for i in 1 ..< count {
                             if v[count - i - 1] > 0 {
                                 let point = CGPoint(x: Double(i) * xScale + 30.0, y: height - Double(v[count - i - 1]) * yScale)
@@ -106,16 +106,16 @@ struct Graph: View {
             // Frame and historic OOP values
             GeometryReader { geometry in
                 Path() { path in
-                    let width  = Double(geometry.size.width) - 60.0
-                    let height = Double(geometry.size.height)
-                    path.addRoundedRect(in: CGRect(x: 0.0 + 30, y: 0.0, width: width, height: height), cornerSize: CGSize(width: 8, height: 8))
+                    let width  = geometry.size.width - 60
+                    let height = geometry.size.height
+                    path.addRoundedRect(in: CGRect(x: 0 + 30, y: 0, width: width, height: height), cornerSize: CGSize(width: 8, height: 8))
                     let count = history.values.count
                     if count > 0 {
                         let v = history.values.map{$0.value}
-                        let yScale = (height - 20.0) / yMax()
+                        let yScale = (height - 20) / yMax()
                         let xScale = width / Double(count - 1)
                         var startingVoid = v[count - 1] < 1 ? true : false
-                        if startingVoid == false { path.move(to: .init(x: 0.0 + 30.0, y: height - Double(v[count - 1]) * yScale)) }
+                        if startingVoid == false { path.move(to: .init(x: 0 + 30, y: height - Double(v[count - 1]) * yScale)) }
                         for i in 1 ..< count {
                             if v[count - i - 1] > 0 {
                                 let point = CGPoint(x: Double(i) * xScale + 30.0, y: height - Double(v[count - i - 1]) * yScale)
